# backend/src/main/java/com/example/walkservice/commerce/dto/UpdateCartItemRequest.java
package com.example.walkservice.commerce.dto;

import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;

public class UpdateCartItemRequest {

    @NotNull
    @Min(1)
    @Max(99)
    private final Integer quantity;

    public UpdateCartItemRequest(Integer quantity) {
        this.quantity = quantity;
    }

    public Integer getQuantity() {
        return quantity;
    }
}